requirements.txt
# requirements.txt

annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
ffmpeg==1.4
filelock==3.15.3
fsspec==2024.6.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==2.1.5
mkl==2021.4.0
more-itertools==10.3.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.0
openai==1.35.3
openai-whisper==20231117
pydantic==2.7.4
pydantic_core==2.18.4
pydub==0.25.1
python-dotenv==1.0.1
regex==2024.5.15
requests==2.32.3
sniffio==1.3.1
sympy==1.12.1
tbb==2021.13.0
tiktoken==0.7.0
torch==2.3.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
